Fix 32-bit builds and the bugs they exposed
The x86 configuration had rotted: it would not compile, and with the
warnings silenced it would not link. Dropping it was the alternative,
but x86 is the cheapest detector this codebase has for size_t/u64
conflation, and an allocator mixing fixed-width bitmaps with
pointer-width sizes is exactly the kind that benefits. It now builds
clean under /W4 /WX and passes every suite in debug and release.

32-bit MSVC provides only the 32-bit bit-scan intrinsics; the 64-bit
forms exist solely on 64-bit targets. A new bitscan_windows.h presents
one spelling to callers, composing two 32-bit scans where the wider
intrinsic is absent. This was the only hard blocker. Everything else
began as a warning.

The arena's size classes assumed a 64-bit address space. The count of
small bins is the span from the smallest chunk to the large-bin
boundary divided by the alignment, which on x86 came to 126 bins
addressed by a 64-bit map: undefined behavior for any request over
roughly 512 bytes. That boundary now derives from the alignment, so
either width gets 62 bins, and the large-bin count is capped by what a
size_t can represent. Every one of these constants evaluates on x64 to
exactly what it did before.

Three defects surfaced that were latent rather than x86-specific:

  - The digital search tree discriminates on the most significant bit
    of a 64-bit word. A 32-bit key never reaches it, so every descent
    went left and the tree degenerated into a list.
  - ALIGN_UP and ALIGN_DOWN built their mask at 64 bits, retyping the
    whole expression so that aligning a size_t produced a value the
    caller had to narrow again.
  - An arena test asserted against a bin left over from an earlier
    loop, so the assertion passed without testing anything.

Tests that baked in 64-bit quantities - an 8 GB arena, a 1 PiB
request, a 64 TiB reservation - now scale to the address space of the
target instead of truncating into something meaningless.

The new header ships in the memory_service and fault_memory_service
packages, which carry the header that includes it. Omitting it also
broke the file and stream packages, which layer over memory_service.

Separately, the stream package's self-test read back what it wrote
through the file service, which that package does not ship, and cast
an opaque handle to a Win32 HANDLE to do it. It now uses stdio and
assumes nothing about the provider's representation.

+++ src/arena.c
@@ -380,7 +380,7 @@ static Chunk *chunk_from_fast(Arena *arena, size_t size, char const *file, int l
 static Chunk *allocate_from_next_nonempty_small_bin(Arena *arena, size_t size,
                                                     char const *file, int line) {
     flag64 small_bits = ARENA_SMALL_BITS(arena, size);
-    u64    index      = ARENA_SMALL_INDEX(size);
+    u32    index      = ARENA_SMALL_INDEX(size);
     flag64 left_bits  = (small_bits << index) & BIT_MASK_LEFT(IDX_TO_BIT64(index));
     flag64 least_bit  = BIT_LSB(left_bits);
     COMPUTE_LSB2IDX64(least_bit, index);
